Cypress component testing *almost* working
